Second version - made it my own
Went back through and took notes on the tutorial. Then, I used the notes to create this app, and I understood it much better. I understand how the data is passed between the 3 modules, and why it's beneficial to structure a project in this way. Also, I added a nicer UI to the top that I like better.